Improved #23 - difficulty levels locked
Addition of a lock for difficulty levels. We can now force the player to not choose certain difficulties. For example, we can force the player to play in easy mode first, and if only he is successful, he can unlock medium and infinity levels

Correction of a small display bug of the score label in G1