Text Alignment for G1 fixed #34 + Small Mod G3 + scale of resized textures
For Game1 - The text is now correctly aligned below the falling virus in the center. I added new constants for Height and Width of small character
    Added debug methods that might be useful in the future
Rescaling of textures that were resized (Mainly in the cutscenes and G1) to make it look like before. We hope it will improve perfs
For Game3 - Small addition of safe/unsafe practices for demo reasons (they are of course not final)