#48 Almost refractored
The cutscene of Game3 is almost finished, new mireille faces expressions has been added as well as new dialogs.
A small bug has also been corrected in the abstract class of cinematic screens, image widths and heights were automatically set to the world width and height, which is incoherent since not all images are the size of the screen, thus this could have led to other problems.