Localization done (it just need to be translated now)
Addition of a button in Option Screen to change the language.
Also rewrote how Localization class works, now the Settings class is responsible to call Localization changeLanguage method. It is done so we also save as a preference the actual language used. Tried to remove some incoherence too about that subject, the use of Java.Locale instead of enum Languages still a bit confuse at times. The enum is preferred to use internally. However, we try to save locale info directly. We use a function to transform a Java locale into a Languages enum.